Fix references always deleted on exit

+++ Tachyon.Game/Database/ArchiveModelManager.cs
@@ -4,10 +4,8 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Humanizer;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore;
-using osu.Framework;
 using osu.Framework.Extensions;
 using osu.Framework.Extensions.IEnumerableExtensions;
 using osu.Framework.Logging;
@@ -22,7 +20,7 @@ using Tachyon.Game.Utils;
 namespace Tachyon.Game.Database
 {
     public abstract class ArchiveModelManager<TModel, TFileModel> : ICanAcceptFiles, IModelManager<TModel>
-        where TModel : class, IHasFiles<TFileModel>, IHasPrimaryKey
+        where TModel : class, IHasFiles<TFileModel>, IHasPrimaryKey, ISoftDelete
         where TFileModel : class, INamedFileInfo, new()
     {
         private static readonly ThreadedTaskScheduler import_scheduler = new ThreadedTaskScheduler(1, nameof(ArchiveModelManager<TModel, TFileModel>));
@@ -32,12 +30,11 @@ namespace Tachyon.Game.Database
         
         public virtual string[] HandledExtensions => new[] { ".zip" };
 
-        public virtual bool SupportsImportFromStable => RuntimeInfo.IsDesktop;
-
         protected readonly FileStore Files;
         protected readonly IDatabaseContextFactory ContextFactory;
         protected readonly MutableDatabaseBackedStore<TModel> ModelStore;
 
+        // ReSharper disable once NotAccessedField.Local (we should keep a reference to this so it is not finalised)
         private ArchiveImportIPCChannel ipc;
 
         protected ArchiveModelManager(Storage storage, IDatabaseContextFactory contextFactory, MutableDatabaseBackedStoreWithFileIncludes<TModel, TFileModel> modelStore, IIpcHost importHost = null)
@@ -93,12 +90,6 @@ namespace Tachyon.Game.Database
         
         protected virtual bool ShouldDeleteArchive(string path) => false;
 
-        /// <summary>
-        /// Import one <typeparamref name="TModel"/> from the filesystem and delete the file on success.
-        /// </summary>
-        /// <param name="path">The archive location on disk.</param>
-        /// <param name="cancellationToken">An optional cancellation token.</param>
-        /// <returns>The imported model, if successful.</returns>
         public async Task<TModel> Import(string path, CancellationToken cancellationToken = default)
         {
             cancellationToken.ThrowIfCancellationRequested();
@@ -107,10 +98,6 @@ namespace Tachyon.Game.Database
             using (ArchiveReader reader = getReaderFrom(path))
                 import = await Import(reader, cancellationToken);
 
-            // We may or may not want to delete the file depending on where it is stored.
-            //  e.g. reconstructing/repairing database with items from default storage.
-            // Also, not always a single file, i.e. for LegacyFilesystemReader
-            // TODO: Add a check to prevent files from storage to be deleted.
             try
             {
                 if (import != null && File.Exists(path) && ShouldDeleteArchive(path))
@@ -124,16 +111,6 @@ namespace Tachyon.Game.Database
             return import;
         }
 
-        /// <summary>
-        /// Fired when the user requests to view the resulting import.
-        /// </summary>
-        public Action<IEnumerable<TModel>> PresentImport;
-
-        /// <summary>
-        /// Import an item from an <see cref="ArchiveReader"/>.
-        /// </summary>
-        /// <param name="archive">The archive to be imported.</param>
-        /// <param name="cancellationToken">An optional cancellation token.</param>
         public Task<TModel> Import(ArchiveReader archive, CancellationToken cancellationToken = default)
         {
             cancellationToken.ThrowIfCancellationRequested();
@@ -160,11 +137,6 @@ namespace Tachyon.Game.Database
             return Import(model, archive, cancellationToken);
         }
 
-        /// <summary>
-        /// Any file extensions which should be included in hash creation.
-        /// Generally should include all file types which determine the file's uniqueness.
-        /// Large files should be avoided if possible.
-        /// </summary>
         protected abstract string[] HashableFileTypes { get; }
 
         protected static void LogForModel(TModel model, string message, Exception e = null)
@@ -177,15 +149,8 @@ namespace Tachyon.Game.Database
                 Logger.Log($"{prefix} {message}", LoggingTarget.Database);
         }
 
-        /// <summary>
-        /// Create a SHA-2 hash from the provided archive based on file content of all files matching <see cref="HashableFileTypes"/>.
-        /// </summary>
-        /// <remarks>
-        ///  In the case of no matching files, a hash will be generated from the passed archive's <see cref="ArchiveReader.Name"/>.
-        /// </remarks>
         private string computeHash(TModel item, ArchiveReader reader = null)
         {
-            // for now, concatenate all .osu files in the set to create a unique hash.
             MemoryStream hashable = new MemoryStream();
 
             foreach (TFileModel file in item.Files.Where(f => HashableFileTypes.Any(f.Filename.EndsWith)))
@@ -203,12 +168,6 @@ namespace Tachyon.Game.Database
             return item.Hash;
         }
 
-        /// <summary>
-        /// Import an item from a <typeparamref name="TModel"/>.
-        /// </summary>
-        /// <param name="item">The model to be imported.</param>
-        /// <param name="archive">An optional archive to use for model population.</param>
-        /// <param name="cancellationToken">An optional cancellation token.</param>
         public async Task<TModel> Import(TModel item, ArchiveReader archive = null, CancellationToken cancellationToken = default) => await Task.Factory.StartNew(async () =>
         {
             cancellationToken.ThrowIfCancellationRequested();
@@ -219,7 +178,6 @@ namespace Tachyon.Game.Database
             {
                 if (!Delete(item))
                 {
-                    // We may have not yet added the model to the underlying table, but should still clean up files.
                     LogForModel(item, "Dereferencing files for incomplete import.");
                     Files.Dereference(item.Files.Select(f => f.FileInfo).ToArray());
                 }
@@ -234,7 +192,7 @@ namespace Tachyon.Game.Database
 
                 await Populate(item, archive, cancellationToken);
 
-                using (var write = ContextFactory.GetForWrite()) // used to share a context for full import. keep in mind this will block all writes.
+                using (var write = ContextFactory.GetForWrite())
                 {
                     try
                     {
@@ -290,13 +248,10 @@ namespace Tachyon.Game.Database
         {
             using (var usage = ContextFactory.GetForWrite())
             {
-                // Dereference the existing file info, since the file model will be removed.
                 Files.Dereference(file.FileInfo);
 
-                // Remove the file model.
                 usage.Context.Set<TFileModel>().Remove(file);
 
-                // Add the new file info and containing file model.
                 model.Files.Remove(file);
                 model.Files.Add(new TFileModel
                 {
@@ -308,11 +263,6 @@ namespace Tachyon.Game.Database
             }
         }
 
-        /// <summary>
-        /// Perform an update of the specified item.
-        /// TODO: Support file additions/removals.
-        /// </summary>
-        /// <param name="item">The item to update.</param>
         public void Update(TModel item)
         {
             using (ContextFactory.GetForWrite())
@@ -323,12 +273,6 @@ namespace Tachyon.Game.Database
             }
         }
 
-        /// <summary>
-        /// Delete an item from the manager.
-        /// Is a no-op for already deleted items.
-        /// </summary>
-        /// <param name="item">The item to delete.</param>
-        /// <returns>false if no operation was performed</returns>
         public bool Delete(TModel item)
         {
             using (ContextFactory.GetForWrite())
@@ -336,7 +280,7 @@ namespace Tachyon.Game.Database
                 // re-fetch the model on the import context.
                 var foundModel = queryModel().Include(s => s.Files).ThenInclude(f => f.FileInfo).FirstOrDefault(s => s.ID == item.ID);
 
-                if (foundModel == null) return false;
+                if (foundModel == null || foundModel.DeletePending) return false;
 
                 if (ModelStore.Delete(foundModel))
                     Files.Dereference(foundModel.Files.Select(f => f.FileInfo).ToArray());
@@ -344,10 +288,6 @@ namespace Tachyon.Game.Database
             }
         }
 
-        /// <summary>
-        /// Delete multiple items.
-        /// This will post notifications tracking progress.
-        /// </summary>
         public void Delete(List<TModel> items, bool silent = false)
         {
             if (items.Count == 0) return;
@@ -357,12 +297,8 @@ namespace Tachyon.Game.Database
                 Delete(b);
             }
         }
-
-        /// <summary>
-        /// Restore multiple items that were previously deleted.
-        /// This will post notifications tracking progress.
-        /// </summary>
-        public void Undelete(List<TModel> items, bool silent = false)
+        
+        public void Undelete(List<TModel> items)
         {
             if (!items.Any()) return;
 
@@ -372,10 +308,6 @@ namespace Tachyon.Game.Database
             }
         }
 
-        /// <summary>
-        /// Restore an item that was previously deleted. Is a no-op if the item is not in a deleted state, or has its protected flag set.
-        /// </summary>
-        /// <param name="item">The item to restore</param>
         public void Undelete(TModel item)
         {
             using (var usage = ContextFactory.GetForWrite())
@@ -390,9 +322,6 @@ namespace Tachyon.Game.Database
             }
         }
 
-        /// <summary>
-        /// Create all required <see cref="System.IO.FileInfo"/>s for the provided archive, adding them to the global file store.
-        /// </summary>
         private List<TFileModel> createFileInfos(ArchiveReader reader, FileStore files)
         {
             var fileInfos = new List<TFileModel>();
@@ -401,7 +330,6 @@ namespace Tachyon.Game.Database
             if (!(prefix.EndsWith("/") || prefix.EndsWith("\\")))
                 prefix = string.Empty;
 
-            // import files to manager
             foreach (string file in reader.Filenames)
             {
                 using (Stream s = reader.GetStream(file))
@@ -417,56 +345,22 @@ namespace Tachyon.Game.Database
             return fileInfos;
         }
 
-        /// <summary>
-        /// Create a barebones model from the provided archive.
-        /// Actual expensive population should be done in <see cref="Populate"/>; this should just prepare for duplicate checking.
-        /// </summary>
-        /// <param name="archive">The archive to create the model for.</param>
-        /// <returns>A model populated with minimal information. Returning a null will abort importing silently.</returns>
         protected abstract TModel CreateModel(ArchiveReader archive);
 
-        /// <summary>
-        /// Populate the provided model completely from the given archive.
-        /// After this method, the model should be in a state ready to commit to a store.
-        /// </summary>
-        /// <param name="model">The model to populate.</param>
-        /// <param name="archive">The archive to use as a reference for population. May be null.</param>
-        /// <param name="cancellationToken">An optional cancellation token.</param>
         protected virtual Task Populate(TModel model, [CanBeNull] ArchiveReader archive, CancellationToken cancellationToken = default) => Task.CompletedTask;
 
-        /// <summary>
-        /// Perform any final actions before the import to database executes.
-        /// </summary>
-        /// <param name="model">The model prepared for import.</param>
         protected virtual void PreImport(TModel model)
         {
         }
 
-        /// <summary>
-        /// Check whether an existing model already exists for a new import item.
-        /// </summary>
-        /// <param name="model">The new model proposed for import.</param>
-        /// <returns>An existing model which matches the criteria to skip importing, else null.</returns>
         protected TModel CheckForExisting(TModel model) => model.Hash == null ? null : ModelStore.ConsumableItems.FirstOrDefault(b => b.Hash == model.Hash);
 
-        /// <summary>
-        /// After an existing <typeparamref name="TModel"/> is found during an import process, the default behaviour is to restore the existing
-        /// item and skip the import. This method allows changing that behaviour.
-        /// </summary>
-        /// <param name="existing">The existing model.</param>
-        /// <param name="import">The newly imported model.</param>
-        /// <returns>Whether the existing model should be restored and used. Returning false will delete the existing and force a re-import.</returns>
         protected virtual bool CanUndelete(TModel existing, TModel import) => true;
 
         private DbSet<TModel> queryModel() => ContextFactory.Get().Set<TModel>();
 
         protected virtual string HumanisedModelName => $"{typeof(TModel).Name.Replace("Info", "").ToLower()}";
 
-        /// <summary>
-        /// Creates an <see cref="ArchiveReader"/> from a valid storage path.
-        /// </summary>
-        /// <param name="path">A file or folder path resolving the archive content.</param>
-        /// <returns>A reader giving access to the archive's content.</returns>
         private ArchiveReader getReaderFrom(string path)
         {
             if (ZipUtils.IsZipArchive(path))
@@ -483,20 +377,10 @@ namespace Tachyon.Game.Database
 
         private readonly List<Action> queuedEvents = new List<Action>();
 
-        /// <summary>
-        /// Allows delaying of outwards events until an operation is confirmed (at a database level).
-        /// </summary>
         private bool delayingEvents;
-
-        /// <summary>
-        /// Begin delaying outwards events.
-        /// </summary>
+        
         private void delayEvents() => delayingEvents = true;
 
-        /// <summary>
-        /// Flush delayed events and disable delaying.
-        /// </summary>
-        /// <param name="perform">Whether the flushed events should be performed.</param>
         private void flushEvents(bool perform)
         {
             Action[] events;
